Funciona eliminar cantores
Hay que tener en cuenta que para eliminar, no puedo tener ningún objeto que dependa del que quiero eliminar

+++ chorus/app/Http/Controllers/CoroController.php
@@ -109,6 +109,8 @@ class CoroController extends Controller
     public function eliminarCoro($id)
     {
         $coro = Coro::find($id);
-        $coro->delete();
+        if ($coro) {
+            $coro->delete();
+        }
     }
 }
